employee/requirements.txt
asgiref==3.3.1
boto3==1.17.48
botocore==1.20.48
Django==3.1.7
django-ses==2.0.0
django-sms==0.5.0
djangorestframework==3.12.2
future==0.18.2
jmespath==0.10.0
pyotp==2.6.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
